refactor: merge AE2 wireless turtle upgrades
